Rename package: person to flashcard
Utilized refactoring, with some minor checkstyle fixes.
