+++ src/main/java/seedu/address/model/flashcard/QuestionContainsKeywordsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.flashcard;
 
 import java.util.List;
 import java.util.function.Predicate;
