+++ src/main/java/seedu/address/model/flashcard/exceptions/PersonNotFoundException.java
@@ -1,0 +1,6 @@
+package seedu.address.model.flashcard.exceptions;
+
+/**
+ * Signals that the operation is unable to find the specified flashcard.
+ */
+public class PersonNotFoundException extends RuntimeException {}
