+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,9 +6,9 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.person.Answer;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Question;
+import seedu.address.model.flashcard.Answer;
+import seedu.address.model.flashcard.Person;
+import seedu.address.model.flashcard.Question;
 import seedu.address.model.tag.Tag;
 
 /**
