+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -24,9 +24,9 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.AddCommand;
-import seedu.address.model.person.Answer;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Question;
+import seedu.address.model.flashcard.Answer;
+import seedu.address.model.flashcard.Person;
+import seedu.address.model.flashcard.Question;
 import seedu.address.model.tag.Tag;
 import seedu.address.testutil.PersonBuilder;
