+++ src/test/java/seedu/address/storage/JsonAdaptedPersonTest.java
@@ -12,8 +12,8 @@ import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.person.Answer;
-import seedu.address.model.person.Question;
+import seedu.address.model.flashcard.Answer;
+import seedu.address.model.flashcard.Question;
 
 public class JsonAdaptedPersonTest {
     private static final String INVALID_QUESTION = "wh@";
